refactor: Apply review
1. Move package reservation.user (UserReservation)
2. Rename Class name
