+++ src/main/java/com/project/escape/domain/reservation/user/UserReservationRepository.java
@@ -1,4 +1,4 @@
-package com.project.escape.domain.reservation;
+package com.project.escape.domain.reservation.user;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
