+++ src/main/java/com/project/escape/domain/shop/Address.java
@@ -12,7 +12,7 @@ import lombok.Setter;
 @Embeddable
 public class Address {
     @Enumerated(EnumType.STRING)
-    private LocationType location;
+    private Location location;
     @Column(nullable = false)
     private String addressDetail;
 }
