+++ src/main/java/com/project/escape/domain/shop/Location.java
@@ -1,6 +1,6 @@
 package com.project.escape.domain.shop;
 
-public enum LocationType {
+public enum Location {
     SEOUL, INCHEON, DAEGU, SEJONG, DAEJEON, GWANGJU, ULSAN, BUSAN,
     GYEONGGI, GANGWON, CHUNGCHEONG, JEOLLA, GYEONGSANG, JEJU
